backend/requirements.txt
fastapi
uvicorn
pydantic
jinja2
google-generativeai
python-dotenv
